Refactor cloudshell
