+++ onyxia-api/src/main/resources/application.properties
@@ -31,9 +31,9 @@ kubernetes.publish.domain=selfservice-kub.example.com
 catalogs.configuration=classpath:multiverse.json
 catalogs.refresh.ms=300000
 
-# Cloudshell
-cloudshell.catalogid=internal
-cloudshell.packagename=shelly
+# Cloudshell (leave blank for defaults)
+cloudshell.catalogid=
+cloudshell.packagename=
 
 # Debug
 debug.http.log=false
